requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.110.1
gradio==4.26.0
loguru==0.7.2
pandas==2.2.2
pydantic==2.7.0
PyYAML==6.0.1
requests==2.31.0
transformers==4.39.3
uvicorn==0.29.0
